+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 git+https://github.com/gradsflow/gradsflow@main
-mkdocs>=1.2.2
+mkdocs>=1.2.3
 mkdocs-material>=7.2.4
 mkdocs-material-extensions==1.0.1
 mkdocs-git-revision-date-localized-plugin==0.9.2
