+++ gradsflow/callbacks/logger/logger.py
@@ -18,7 +18,7 @@ from typing import Optional
 import pandas as pd
 from loguru import logger
 
-from gradsflow.core.callbacks import Callback
+from gradsflow.callbacks.base import Callback
 from gradsflow.utility.common import to_item
